+++ client/src/constants/Config.js
@@ -1,1 +1,1 @@
-export const API_URL = 'https://5e6ca313d708a000160b5225.mockapi.io';
+export const API_URL = 'http://localhost:8000';
