+++ server/src/app.js
@@ -6,12 +6,14 @@ const errorHandler = require('errorhandler');
 const session = require('express-session');
 const bodyParser = require('body-parser');
 import 'reflect-metadata';
+import cors from "cors";
 
 // routers
 import indexRouter from './routes';
 
 // configs
 import { dataConnection } from './config/database';
+import corsConfig from './config/cors';
 
 const isProduction = process.env.NODE_ENV === 'production';
 
@@ -24,6 +26,19 @@ app.use(bodyParser.urlencoded({ extended: false }));
 app.use(bodyParser.json());
 app.use(session({ secret: 'LightBlog', cookie: { maxAge: 60000 }, resave: false, saveUninitialized: false }));
 
+// setting cors
+const whitelist = corsConfig.CORS;
+const corsOptions = {
+    origin: function (origin, callback) {
+        console.log(whitelist, origin, whitelist.indexOf(origin))
+        if (origin === undefined || whitelist.indexOf(origin) !== -1) {
+            callback(null, true)
+        } else {
+            callback(new Error('Not allowed by CORS'))
+        }
+    }
+}
+app.use(cors(corsOptions));
 
 // path for file static
 app.use(express.static(path.join(__dirname, '../public')));
